+++ main.py
@@ -1,7 +1,6 @@
 import cmd
-import json
 import os
-import re
+import json
 
 
 class wShell(cmd.Cmd):
@@ -16,8 +15,8 @@ class wShell(cmd.Cmd):
     remembered_dirs = [os.getcwd()]
     arithmetic_ops = ["-eq", "-ne", "-lt", "-le", "-gt", "-ge"]
 
-    import built_in
-    # from built_in import alias, cd, dirs, echo, test
+    from built_in import alias, cd, dirs, echo, logic
+    from extras import cat, ls, path
 
     @staticmethod
     def emptyline():
@@ -32,21 +31,6 @@ class wShell(cmd.Cmd):
     def do_shell(line: str):
         return os.system(line)
 
-    @staticmethod
-    def command_in_path(exe: str):
-        """Checks in every address in the system PATH enviroment variable whether the command exists or not. If it does,
-        it executes it. Otherwise, returns False"""
-
-        for path in os.environ['PATH'].split(";"):  # For each value in the PATH
-            if not os.path.isdir(path):  # If the PATH address is not valid, move on
-                continue
-
-            if exe + '.exe' in os.listdir(
-                    path):  # However, if an executable file with the given name exists, return True
-                return True
-
-        return False  # Otherwise, return False (the command is not in PATH)
-
     def postcmd(self, stop, line: str):
         self.prompt = str(0 if stop is None else stop) + "<" + str(os.getcwd()) + ">"
         return
@@ -55,9 +39,9 @@ class wShell(cmd.Cmd):
         args_list = line.split(' ')
         if self.aliases.get(args_list[0], None):
             self.onecmd(self.aliases[args_list[0]])
-        elif self.is_assignment(line):
+        elif logic.is_assignment(line):
             print(self.variables)
-        elif self.command_in_path(args_list[0]) or os.path.isfile(args_list[0]):
+        elif path.command_in_path(args_list[0]) or os.path.isfile(args_list[0]):
             # If the command is a file found in the PATH or the command itself is a file, execute it as is
             return self.do_shell(' '.join(args_list))
         else:
@@ -71,7 +55,7 @@ class wShell(cmd.Cmd):
             stdout.write("Alias file not found, creating...\n")
             with open(self.aliasfile, 'a') as a:
                 a.write(json.dumps(self.aliases))
-        do_reloadalias("")
+        alias.do_reloadalias("")
 
 
 if __name__ == "__main__":
